outsourced utils into new src file and added scala Doc
